requirements.txt
asgiref==3.5.2
asyncio==3.4.3
certifi==2022.6.15
chardet==4.0.0
dj-database-url==0.5.0
Django==4.0.5
django-heroku==0.3.1
djangorestframework==3.13.1
fyers-apiv2==2.0.5
gunicorn==20.1.0
idna==2.10
psycopg2==2.9.3
pytz==2022.1
requests==2.25.1
sqlparse==0.4.2
tornado==6.1
tzdata==2022.1
urllib3==1.26.9
websocket-client==1.2.1
websockets==8.1
whitenoise==6.2.0
